FATAL creates a popup.
A few more changes to the way error handling is handled.

When not in development only errors should be shown in the log, fatal errors will create a red popup.
If ncurses is not active, the error message will be written using printf();

I believe this is all the changes I wish to make to v3. Feel free to finish it up.
I will now begin working on v4, which will feature framebuffer graphics :)